Move "upload" detection logic into separate resolver to cleanup source generator.
